requirements.txt
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.23.1
pymongo==3.12.3
python-dotenv==0.19.2
Werkzeug==2.0.2
WTForms==3.0.1
